worker/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
Pillow==10.2.0
numpy==1.26.4
numba==0.59.0
